+++ cmd/tools/cvtgxl.c
@@ -54,7 +54,7 @@ static FILE *getFile(void)
 }
 #endif
 
-static FILE *openFile(char *name, char *mode)
+static FILE *openFile(const char *name, const char *mode)
 {
     FILE *fp;
     char *modestr;