+++ doc/libgraph/sccmap.c
@@ -252,7 +252,7 @@ static void process(Agraph_t * G)
 
 }
 
-static FILE *openFile(char *name, char *mode)
+static FILE *openFile(const char *name, const char *mode)
 {
 	FILE *fp;
 	char *modestr;