+++ lib/label/nrtmain.c
@@ -223,7 +223,7 @@ void usage(char *pp)
     return;
 }
 
-static FILE *openFile(char *name, char *mode)
+static FILE *openFile(const char *name, const char *mode)
 {
     FILE *fp;
     char *modestr;